// needrelay-frontend/src/api/generated/models/reliefRequestResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { ReliefRequestResponseStatus } from './reliefRequestResponseStatus';

export interface ReliefRequestResponse {
  id: string;
  title: string;
  description: string;
  locationLabel: string;
  latitude: number;
  longitude: number;
  publicSlug: string;
  status: ReliefRequestResponseStatus;
  createdAt: string;
}