// needrelay-frontend/src/api/generated/models/needCategory.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */

export type NeedCategory = typeof NeedCategory[keyof typeof NeedCategory];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NeedCategory = {
  FOOD: 'FOOD',
  WATER: 'WATER',
  CLOTHING: 'CLOTHING',
  SUPPLIES: 'SUPPLIES',
  EQUIPMENT: 'EQUIPMENT',
  MACHINERY: 'MACHINERY',
  VEHICLES: 'VEHICLES',
  SHELTER: 'SHELTER',
  SERVICES: 'SERVICES',
  OTHER: 'OTHER',
} as const;